requirements.txt
dotenv
psycopg2-binary
Faker
langchain
openai
tiktoken
langchain-openai
langchain-community
pandas
streamlit
fastapi
uvicorn
tqdm
pypdf
faiss-cpu